move memory subsystem stuff into a seperate module
