+++ kernel/src/main.rs
@@ -10,74 +10,16 @@ extern crate alloc;
 
 core::arch::global_asm!(core::include_str!("./start.S"));
 
+mod memory;
 mod running_image;
 mod uart;
 
 use core::fmt::Write;
 
-use itertools::Itertools;
 use kernel_core::{
-    memory::{BuddyPageAllocator, HeapAllocator, PhysicalPointer},
+    memory::{PageAllocator, PhysicalPointer},
     platform::device_tree::{DeviceTree, Value as DTValue},
 };
-
-use spin::once::Once;
-
-/// The global physical page allocator.
-static PAGE_ALLOCATOR: Once<BuddyPageAllocator> = Once::new();
-
-#[global_allocator]
-/// The Rust global heap allocator.
-static ALLOCATOR: HeapAllocator<'static, BuddyPageAllocator> = HeapAllocator::new_uninit();
-
-/// Initialize the memory subsystem.
-///
-/// The memory subsystem consists of:
-/// - the global physical page allocator
-/// - the MMU and the kernel page tables
-/// - the Rust heap
-fn init_memory(dt: &DeviceTree<'_>, uart: &mut impl Write) {
-    // create page allocator
-    let page_size = 0x1000;
-    let memory_node = dt
-        .iter_nodes_named(b"/", b"memory")
-        .expect("root")
-        .exactly_one()
-        .expect("device tree has memory node");
-    let memory_range = memory_node
-        .properties
-        .clone()
-        .find(|(name, _)| name == b"reg")
-        .and_then(|(_, v)| v.into_reg())
-        .expect("memory node has reg property")
-        .iter()
-        .exactly_one()
-        .expect("memory has exactly one reg range");
-    let reserved_regions = [
-        unsafe { running_image::memory_region() },
-        dt.memory_region(),
-    ];
-    writeln!(
-        uart,
-        "memory range = {memory_range:x?}, reserved = {reserved_regions:x?}"
-    )
-    .unwrap();
-    let pa = PAGE_ALLOCATOR.call_once(|| unsafe {
-        BuddyPageAllocator::new(
-            page_size,
-            PhysicalPointer::from(memory_range.0).into(),
-            memory_range.1,
-            reserved_regions.into_iter(),
-        )
-    });
-
-    // setup page tables
-
-    // initialize kernel heap
-    ALLOCATOR.init(pa);
-
-    writeln!(uart, "memory initialized").unwrap();
-}
 
 /// The main entry point for the kernel.
 ///
@@ -125,7 +67,14 @@ pub extern "C" fn kmain(device_tree_blob: PhysicalPointer<u8>) -> ! {
     })
     .unwrap();
 
-    init_memory(&device_tree, &mut uart);
+    memory::init(&device_tree, &mut uart);
+
+    writeln!(
+        &mut uart,
+        "page size = {}",
+        memory::page_allocator().page_size()
+    )
+    .unwrap();
 
     #[allow(clippy::empty_loop)]
     loop {}
