+++ kernel_core/src/memory/buddy.rs
@@ -45,7 +45,7 @@ impl<const MAX_ORDER: usize> BuddyPageAllocator<MAX_ORDER> {
     ///
     /// The memory region provided must be entirely valid memory that is safe to dereference, live for the lifetime of the allocator and not be shared
     /// outside of the allocator.
-    /// Additionally the `excluded_regions` must be contained within the overall memory region, see [`subtract_ranges`] for details.
+    /// Additionally the `excluded_regions` must be contained within the overall memory region, see [`subtract_ranges()`] for details.
     pub unsafe fn new(
         page_size: usize,
         memory_start: *mut u8,
