+++ kernel_core/src/memory/mod.rs
@@ -9,8 +9,8 @@ pub use buddy::BuddyPageAllocator;
 mod heap;
 pub use heap::HeapAllocator;
 
-pub mod subtract_ranges;
-pub use subtract_ranges::subtract_ranges;
+mod subtract_ranges;
+pub use subtract_ranges::*;
 
 /// A 48-bit physical address pointer that is not part of a virtual address space.
 ///
